Limit matrix crack alpha to non-delam orientations
Previously, the orientation of matrix cracks evaluated by DGD.for would span the full range of orienations at a set interval. Because the intended use of CompDam is to have layers of cohesive elements to represent delaminations between differently-oriented solid-element plies, allowing for matrix cracks to be inserted with delamination-like orientations is redundant.