# Updated Params file to have schaefer vals (#11)
* # Updated Params *params file to have schaefer vals

Values for convergence, for schaefers module, were previously hardcoded in the
schafer module. Two new parameters were added:

schaefer_nr_tolerance
schaefer_nr_counter_limit.

These directly correspond to the previously hardcoded values, scoped to schaefer_mod.

In addition to adding these files to CompDam.parameters, the parameters.for
type was update to have places for these values to be saved. Additionally,
the variable p is now passed to plasticity so that it may be used by schaefer.

* # Somehow had removed a portion of code which is in current dev/master

In dgd.for theere is a check that wasn't in my branch. Re-add it in

* # Removed line from test_runner which was specific to my machine