Added fiber nonlinearity to bilinear fiber failure
Fiber nonlinearity was introduced with the fiber kinking model, but it
was not properly integrated with the existing traditional CDM model for
fiber damage (bilinear softening).

This commit allows for fiber nonlinearity to be used with the CDM model
for fiber failure. Before damage, the fiber direction stiffness has an
elastic nonlinearity defined by the material input property cl. After
damage initiation, the secant stiffness at the damage initiation strain
is used. The energy dissipation is consistent no matter the value of cl.

The maximum value for cl was reduced to help prevent negative tangent
stiffnesses from occuring.

In addition, a bug in the reloading behavior of fiber damage, from
compression to tension, is fixed.