Added fiber failure criterion to FKT model
A new parameter is added (fkt_fiber_failure_angle) that defines the fiber rotation at which the
fibers are assumed to fail, after a kink band has formed. Typical
values are between 5 and 30 degrees. This values is not a well defined
material property, and so it is implemented as a parameter to reflect
the potential for problem dependence.

Improvements to convergence of DGDkinkband:
Added cutback logic to DGDkinkband
Added shear nonlinearity to jacobian for DGDkinkband
Added initializations for some previously uninitialized variables