Fixes for FKT convergence
Implemented a restart for DGDKinkband that uses a modified jacobian

Added terminate_on_no_covergence flag to parameters.for to make the
behavior when no converged solution is found in DGDEvolve and DGDKinkband
user-defined. When set to .TRUE. the analysis will terminate when no
converged solution is found (default). When set to false the element
will be deleted and a warning will be issues when no converged solution
is found.

Updated twoElement_fiberCompression_FKT test to use 26 state variables.