Several improvements to debug.py and pyextmod
Refactored code for writing debug.py so that it's broken up among the
different modules in which the data resides. The intent is to ease
maintainability. The element number is now included in the debug.py file
name along with a counter so that multiple debug.py files can be written
during an analysis. The output format of the state variables and
material properties is switched from a dictionary to simplify reading
the data into the python extension module.

Two bash scripts are added to streamline execution of the python extension
module in conjunction with the debug.py files.

Tests implemented for the debug.py files to ensure that future updates
to CompDam include appropriate updates to the debug.py logic.

Minor improvements to setup.py to check for dependencies needed for use
of the python extension module.