Clean analysis termination
- Use vexternaldb to cleanly terminate analyses at the end of an 
increment.
- New logger function for clean termination (log%terminate). log%error 
is still used for immediate termination of an analysis. log%error should 
still be used for dealing with input errors and numerical errors. 
Convergence errors should usually be able to use log%terminate without 
issue.
- Added a test which causes a DGD convergence issue in a single element. 
With the other changes in this commit, an additional output frame is 
generated. Previously, no data from the last model state would be added 
to the ODB file.